requirements.txt
streamlit>=1.22.0
openai-whisper>=20231117
torch>=2.0.0
torchaudio>=2.0.0
transformers>=4.30.0
datasets>=2.12.0
numpy>=1.24.0
pandas>=2.0.0
pydub>=0.25.1
scipy>=1.10.0
